requirements.txt
Flask
pymongo
PyJWT
bcrypt
python-dotenv
gunicorn